added rest api functionality.
