--- backend/models/product.js
@@ -1,1 +1,0 @@
-import mongoose from 'mongoose';
