+++ backend/models/product.model.js
@@ -2,15 +2,19 @@ import mongoose from "mongoose";
 
 const productSchema = new mongoose.Schema(
 	{
-		name: {
+		/*address: {
+			type: String,
+			required: true,
+		},*/
+		image: {
 			type: String,
 			required: true,
 		},
-		price: {
-			type: Number,
+		severity: {
+			type: String,
 			required: true,
 		},
-		image: {
+		description: {
 			type: String,
 			required: true,
 		},
