+++ backend/routes/product.route.js
@@ -1,30 +1,12 @@
 import express from "express";
-import dotenv from "dotenv";
-import path from "path";
 
-import { connectDB } from "./config/db.js";
+import { createProduct, deleteProduct, getProducts, updateProduct } from "../controllers/product.controller.js";
 
-import productRoutes from "./routes/product.route.js";
+const router = express.Router();
 
-dotenv.config();
+router.get("/", getProducts);
+router.post("/", createProduct);
+router.put("/:id", updateProduct);
+router.delete("/:id", deleteProduct);
 
-const app = express();
-const PORT = process.env.PORT || 5000;
-
-const __dirname = path.resolve();
-
-app.use(express.json()); // allows us to accept JSON data in the req.body
-
-app.use("/api/products", productRoutes);
-
-if (process.env.NODE_ENV === "production") {
-	app.use(express.static(path.join(__dirname, "/frontend/dist")));
-	app.get("*", (req, res) => {
-		res.sendFile(path.resolve(__dirname, "frontend", "dist", "index.html"));
-	});
-}
-
-app.listen(PORT, () => {
-	connectDB();
-	console.log("Server started at http://localhost:" + PORT);
-});
+export default router;
